% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARcov.R
\name{ARcov}
\alias{ARcov}
\title{Function to generate an AR(1) variance-covariance matrix with parameter rho s.t. \eqn{\lvert \rho\rvert  < 1}.}
\usage{
ARcov(rho, n)
}
\arguments{
\item{rho}{Numeric value representing the autocorrelation parameter. Must satisfy |rho| < 1.}

\item{n}{Integer representing the size of the matrix (n x n).}
}
\value{
A variance-covariance matrix of size n x n based on the AR(1) process.
}
\description{
Function to generate an AR(1) variance-covariance matrix with parameter rho s.t. \eqn{\lvert \rho\rvert  < 1}.
}
\keyword{internal}
