% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARcov_lit.R
\name{ARcov_lit}
\alias{ARcov_lit}
\title{Function to generate an ARIMA(1,1,0) variance-covariance matrix for the Litterman method with parameter rho s.t. |rho| < 1.}
\usage{
ARcov_lit(rho, n)
}
\description{
Function to generate an ARIMA(1,1,0) variance-covariance matrix for the Litterman method with parameter rho s.t. |rho| < 1.
}
\keyword{internal}
