% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chowlin_likelihood.R
\name{chowlin_likelihood}
\alias{chowlin_likelihood}
\title{Likelihood function from Chow-Lin or Litterman temporal disaggregation.}
\usage{
chowlin_likelihood(Y, X, vcov)
}
\arguments{
\item{Y}{The low-frequency response series (n_l x 1 matrix).}

\item{X}{The aggregated high-frequency indicator series (n_l x p matrix).}

\item{vcov}{Aggregated variance-covariance matrix of Chow-Lin or Litterman residuals.}
}
\description{
Used in disaggregation.R to find estimates of the optimal rho parameter.
}
\references{
\insertAllCited{}
}
\keyword{chow}
\keyword{disaggregation}
\keyword{lin}
\keyword{litterman}
\keyword{temporal}
