% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptd_BIC.R
\name{sptd_BIC}
\alias{sptd_BIC}
\title{BIC Score for Sparse Temporal Disaggregation}
\usage{
sptd_BIC(Y, X, vcov)
}
\arguments{
\item{Y}{The low-frequency response series (\eqn{n_l \times 1} matrix).}

\item{X}{The aggregated high-frequency indicator series (\eqn{n_l \times p} matrix).}

\item{vcov}{Aggregated variance-covariance matrix of AR(\eqn{1}) residuals (\eqn{n_l \times n_l} matrix).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    The minimum BIC score from the refitted models.
}\if{html}{\out{</div>}}
}
\description{
This function calculates the BIC score for sparse temporal disaggregation, as described in
\insertCite{10.1111/rssa.12952;textual}{DisaggregateTS}. It uses the LARS algorithm to find the
optimal beta coefficients and refits the models to compute BIC scores.
}
\keyword{internal}
