% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARcov.R
\name{ARcov}
\alias{ARcov}
\title{Function to generate an AR(1) variance-covariance matrix with parameter rho s.t. |rho| < 1.}
\usage{
ARcov(rho, n)
}
\description{
Function to generate an AR(1) variance-covariance matrix with parameter rho s.t. |rho| < 1.
}
\keyword{internal}
