% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARcov_lit.R
\name{ARcov_lit}
\alias{ARcov_lit}
\title{Function to generate an ARIMA(1,1,0) variance-covariance matrix for the Litterman method with parameter \eqn{\rho} such that \eqn{\lvert \rho \rvert < 1}.}
\usage{
ARcov_lit(rho, n)
}
\arguments{
\item{rho}{Numeric value representing the autocorrelation parameter. Must satisfy \eqn{\lvert \rho \rvert < 1}.}

\item{n}{Integer representing the size of the matrix \eqn{(n x n)}.}
}
\value{
A variance-covariance matrix of size \eqn{(n x n)} for the ARIMA(1,1,0) process, used in the Litterman method.
}
\description{
Function to generate an ARIMA(1,1,0) variance-covariance matrix for the Litterman method with parameter \eqn{\rho} such that \eqn{\lvert \rho \rvert < 1}.
}
\keyword{internal}
