% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdBIC.R
\name{hdBIC}
\alias{hdBIC}
\title{BIC score}
\usage{
hdBIC(X, Y, covariance, beta)
}
\arguments{
\item{X}{Aggregated indicator series matrix that has been GLS rotated.}

\item{Y}{Low-frequency response vector that has been GLS rotated.}

\item{covariance}{Aggregated AR covariance matrix.}

\item{beta}{Estimate of beta from LARS algorithm for a certain lambda.}
}
\description{
This function calculates the BIC score that has been shown to work better than ordinary BIC in
high-dimensional scenarios. It uses the variance estimator given in \insertCite{yu2019estimating;textual}{DisaggregateTS}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
