% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindex.R
\name{k.index}
\alias{k.index}
\title{Index of support for LARS algorithm when in high-dimensions}
\usage{
k.index(matrix, n_l)
}
\description{
This function prevents the support of beta becoming greater than n_l/2.
This heuristic approach prevents erratic values of BIC when in high-dimensions.
}
\keyword{internal}
