% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.R
\name{refit}
\alias{refit}
\title{Refit LASSO estimate into GLS}
\usage{
refit(X, Y, beta)
}
\arguments{
\item{X}{Aggregated indicator series matrix that has been GLS rotated.}

\item{Y}{Low-frequency response vector that has been GLS rotated.}

\item{beta}{Estimate of beta from LARS algorithm for a certain lambda.}
}
\description{
This function reduces the bias in LASSO estimates by re-fitting the
support back into GLS.
}
\keyword{internal}
