% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptd_BIC.R
\name{sptd_BIC}
\alias{sptd_BIC}
\title{Function to calculate the BIC score from sparse temporal disaggregation.}
\usage{
sptd_BIC(Y, X, vcov)
}
\arguments{
\item{Y}{The low-frequency response series (n_l x 1 matrix).}

\item{X}{The aggregated high-frequency indicator series (n_l x p matrix).}

\item{vcov}{Aggregated variance-covariance matrix of AR(1) residuals.}
}
\description{
Used in disaggregation.R to find estimates of the optimal rho parameter.
}
\references{
\insertAllCited{}
}
\keyword{chow}
\keyword{disaggregation}
\keyword{lin}
\keyword{litterman}
\keyword{temporal}
