% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempDisaggToolbox.R
\name{TempDisaggToolbox}
\alias{TempDisaggToolbox}
\title{Low-dimensional temporal dissaggregation toolbox}
\usage{
TempDisaggToolbox(
  Y,
  X = matrix(data = rep(1, times = nrow(Y)), nrow = nrow(Y)),
  method = "Denton-Cholette",
  annualMat = "sum",
  Denton = "first"
)
}
\arguments{
\item{Y}{the low-frequency response vector}

\item{X}{the high-frequency indicator matrix}

\item{method}{nominates the choice of the temporal disaggregation method}

\item{annualMat}{choice of the disaggregation matrix}

\item{Denton}{absolute, first, second and proportional difference Sigma for the Denton method}
}
\description{
This function contains the movement preservation and regression-based low-dimensional temporal disaggregation methods proposed by \insertCite{denton1971adjustment;textual}{DisaggregateTS}, \insertCite{dagum2006benchmarking;textual}{DisaggregateTS}
\insertCite{chow1971best;textual}{DisaggregateTS}, \insertCite{fernandez1981methodological;textual}{DisaggregateTS} and \insertCite{litterman1983random;textual}{DisaggregateTS}.
}
\examples{
TempDisaggToolbox(Y = Y_Gen, X = X_Gen, method = 'Chow-Lin', annualMat = 'sum')
}
\references{
\insertAllCited{}
}
\keyword{Chow-Lin}
\keyword{Denton}
\keyword{Denton-Cholette}
\keyword{Fernandez}
\keyword{Litterman}
\keyword{temporal-disaggregation}
